Move @Entity classes to their own package.
