+++ app/src/main/java/uk/me/jeremygreen/merging2/model/CoordinateDao.kt
@@ -3,6 +3,7 @@ package uk.me.jeremygreen.merging2.model
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import uk.me.jeremygreen.merging2.model.entity.Coordinate
 
 @Dao
 internal interface CoordinateDao {
