+++ app/src/main/java/uk/me/jeremygreen/merging2/model/FaceWithCoordinates.kt
@@ -1,6 +1,7 @@
 package uk.me.jeremygreen.merging2.model
 
 import androidx.room.Relation
+import uk.me.jeremygreen.merging2.model.entity.Coordinate
 
 internal data class FaceWithCoordinates(
